- Algorithm to put in stage the power_up with their own chance to be spawned.

+++ ProjetPang/Assets/UniqueBonusProbalitySpawn.cs
@@ -3,5 +3,5 @@ using System.Collections;
 
 public abstract class UniqueBonusProbalitySpawn : MonoBehaviour {
 
-	public int chanceToSpawn;
+	public float chanceToSpawn;
 }
